+++ valiu-node-commons/src/asset.rs
